+++ gcovr/utils.py
@@ -17,10 +17,13 @@
 # ****************************************************************************
 
 from argparse import ArgumentTypeError
+import logging
 import os
 import re
 import sys
 from contextlib import contextmanager
+
+logger = logging.getLogger("gcovr")
 
 
 class LoopChecker(object):
@@ -60,6 +63,10 @@ else:
     realpath = os.path.realpath
 
 
+def get_os_independent_path(path):
+    return path.replace(os.path.sep, '/')
+
+
 def search_file(predicate, path, exclude_dirs):
     """
     Given a search path, recursively descend to find files that satisfy a
@@ -80,11 +87,11 @@ def search_file(predicate, path, exclude_dirs):
         dirs[:] = [d for d in dirs
                    if not any(exc.match(os.path.join(root, d))
                               for exc in exclude_dirs)]
-        root = realpath(root)
+        root = os.path.abspath(root)
 
         for name in files:
             if predicate(name):
-                yield realpath(os.path.join(root, name))
+                yield os.path.abspath(os.path.join(root, name))
 
 
 def commonpath(files):
@@ -173,6 +180,18 @@ def get_global_stats(covdata):
             branches_total, branches_covered, percent_branches)
 
 
+def summarize_file_coverage(coverage, root_filter):
+    filename = presentable_filename(
+        coverage.filename, root_filter=root_filter)
+
+    branch_total, branch_covered, branch_percent = coverage.branch_coverage()
+    line_total, line_covered, line_percent = coverage.line_coverage()
+    function_total, function_covered, function_percent = coverage.function_coverage()
+    return (filename, line_total, line_covered, line_percent,
+            branch_total, branch_covered, branch_percent,
+            function_total, function_covered, function_percent)
+
+
 def calculate_coverage(covered, total, nan_value=0.0):
     coverage = nan_value
     if total != 0:
@@ -189,7 +208,7 @@ class FilterOption(object):
         self.regex = regex
         self.path_context = os.getcwd() if path_context is None else path_context
 
-    def build_filter(self, logger):
+    def build_filter(self):
         # Try to detect unintended backslashes and warn.
         # Later, the regex engine may or may not raise a syntax error.
         # An unintended backslash is a literal backslash r"\\",
@@ -197,9 +216,9 @@ class FilterOption(object):
         (suggestion, bs_count) = re.subn(
             r'\\\\|\\(?=[^\WabfnrtuUvx0-9AbBdDsSwWZ])', '/', self.regex)
         if bs_count:
-            logger.warn("filters must use forward slashes as path separators")
-            logger.warn("your filter : {}", self.regex)
-            logger.warn("did you mean: {}", suggestion)
+            logger.warning("filters must use forward slashes as path separators")
+            logger.warning(f"your filter : {self.regex}")
+            logger.warning(f"did you mean: {suggestion}")
 
         isabs = self.regex.startswith("/")
         if not isabs and (sys.platform == "win32"):
@@ -232,7 +251,7 @@ class Filter(object):
         self.pattern = re.compile(pattern, flags)
 
     def match(self, path):
-        os_independent_path = path.replace(os.path.sep, '/')
+        os_independent_path = get_os_independent_path(path)
         return self.pattern.match(os_independent_path)
 
     def __str__(self):
@@ -242,27 +261,27 @@ class Filter(object):
 
 class AbsoluteFilter(Filter):
     def match(self, path):
-        abspath = realpath(path)
-        return super(AbsoluteFilter, self).match(abspath)
+        path = realpath(path)
+        return super(AbsoluteFilter, self).match(path)
 
 
 class RelativeFilter(Filter):
     def __init__(self, root, pattern):
         super(RelativeFilter, self).__init__(pattern)
-        self.root = root
+        self.root = realpath(root)
 
     def match(self, path):
-        abspath = realpath(path)
+        path = realpath(path)
 
         # On Windows, a relative path can never cross drive boundaries.
         # If so, the relative filter cannot match.
         if sys.platform == 'win32':
-            path_drive, _ = os.path.splitdrive(abspath)
-            root_drive, _ = os.path.splitdrive(realpath(self.root))
+            path_drive, _ = os.path.splitdrive(path)
+            root_drive, _ = os.path.splitdrive(self.root)
             if path_drive != root_drive:
                 return None
 
-        relpath = os.path.relpath(abspath, self.root)
+        relpath = os.path.relpath(path, self.root)
         return super(RelativeFilter, self).match(relpath)
 
     def __str__(self):
@@ -280,54 +299,27 @@ class AlwaysMatchFilter(Filter):
 
 class DirectoryPrefixFilter(Filter):
     def __init__(self, directory):
-        abspath = os.path.realpath(directory)
-        os_independent_dir = abspath.replace(os.path.sep, '/')
-        pattern = re.escape(os_independent_dir + '/')
+        path = realpath(directory)
+        os_independent_path = get_os_independent_path(path)
+        pattern = re.escape(f"{os_independent_path}/")
         super(DirectoryPrefixFilter, self).__init__(pattern)
 
     def match(self, path):
-        normpath = os.path.normpath(path)
-        return super(DirectoryPrefixFilter, self).match(normpath)
+        realpath = os.path.normpath(path)
+        return super(DirectoryPrefixFilter, self).match(realpath)
 
 
-class Logger(object):
-    def __init__(self, verbose=False):
-        self.verbose = verbose
+def configure_logging() -> None:
+    logging.basicConfig(
+        format="(%(levelname)s) %(message)s",
+        stream=sys.stderr,
+        level=logging.INFO,
+    )
 
-    def warn(self, pattern, *args, **kwargs):
-        """Write a formatted warning to STDERR.
+    def exception_hook(exc_type, exc_value, exc_traceback) -> None:
+        logging.exception("Uncaught EXCEPTION", exc_info=(exc_type, exc_value, exc_traceback))
 
-        pattern: a str.format pattern
-        args, kwargs: str.format arguments
-        """
-        pattern = "(WARNING) " + pattern + "\n"
-        sys.stderr.write(pattern.format(*args, **kwargs))
-
-    def error(self, pattern, *args, **kwargs):
-        """Write a formatted error to STDERR.
-
-        pattern: a str.format pattern
-        args, kwargs: str.format parameters
-        """
-        pattern = "(ERROR) " + pattern + "\n"
-        sys.stderr.write(pattern.format(*args, **kwargs))
-
-    def msg(self, pattern, *args, **kwargs):
-        """Write a formatted message to STDOUT.
-
-        pattern: a str.format pattern
-        args, kwargs: str.format arguments
-        """
-        pattern = pattern + "\n"
-        sys.stdout.write(pattern.format(*args, **kwargs))
-
-    def verbose_msg(self, pattern, *args, **kwargs):
-        """Write a formatted message to STDOUT if in verbose mode.
-
-        see: self.msg()
-        """
-        if self.verbose:
-            self.msg(pattern, *args, **kwargs)
+    sys.excepthook = exception_hook
 
 
 def sort_coverage(covdata, show_branch,
@@ -432,20 +424,3 @@ def presentable_filename(filename, root_filter):
         normalized = filename
 
     return normalized.replace('\\', '/')
-
-
-def fixup_percent(percent):
-    # output csv percent values in range [0,1.0]
-    return percent / 100 if percent is not None else None
-
-
-def summarize_file_coverage(coverage, root_filter):
-    filename = presentable_filename(
-        coverage.filename, root_filter=root_filter)
-
-    branch_total, branch_covered, branch_percent = coverage.branch_coverage()
-    line_total, line_covered, line_percent = coverage.line_coverage()
-    function_total, function_covered, function_percent = coverage.function_coverage()
-    return (filename, line_total, line_covered, fixup_percent(line_percent),
-            branch_total, branch_covered, fixup_percent(branch_percent),
-            function_total, function_covered, fixup_percent(function_percent))